# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StreamGuard API - Batch IPTV stream health checker and playlist manager
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /playlists:
    get:
      operationId: listPlaylists
      tags: [playlists]
      summary: List all playlists
      responses:
        "200":
          description: List of playlists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Playlist"
    post:
      operationId: createPlaylist
      tags: [playlists]
      summary: Create playlist from text, URL, or file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistInput"
      responses:
        "201":
          description: Playlist created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /playlists/{id}:
    get:
      operationId: getPlaylist
      tags: [playlists]
      summary: Get playlist details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Playlist details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deletePlaylist
      tags: [playlists]
      summary: Delete a playlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /playlists/{id}/channels:
    get:
      operationId: getPlaylistChannels
      tags: [playlists]
      summary: Get channels from a playlist with pagination
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: group
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated channel list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelPage"

  # ── Jobs ─────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Playlist:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sourceType:
          type: string
          enum: [text, url, file]
        sourceUrl:
          type: ["string", "null"]
        entryCount:
          type: integer
        duplicatesFound:
          type: integer
        parseWarnings:
          type: array
          items:
            type: string
        groups:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, sourceType, entryCount, duplicatesFound, parseWarnings, groups, createdAt]
    PlaylistInput:
      type: object
      properties:
        name:
          type: string
        sourceType:
          type: string
          enum: [text, url, file]
        content:
          type: ["string", "null"]
          description: Raw M3U/M3U8 text (for sourceType=text) or base64 file (for sourceType=file)
        url:
          type: ["string", "null"]
          description: Remote URL to fetch (for sourceType=url)
      required: [name, sourceType]
    Channel:
      type: object
      properties:
        id:
          type: integer
        playlistId:
          type: integer
        tvgId:
          type: ["string", "null"]
        tvgName:
          type: ["string", "null"]
        tvgLogo:
          type: ["string", "null"]
        groupTitle:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        url:
          type: string
        position:
          type: integer
      required: [id, playlistId, url, position]
    ChannelPage:
      type: object
      properties:
        channels:
          type: array
          items:
            $ref: "#/components/schemas/Channel"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [channels, total, page, limit]

    # ── Job schemas ─────────────────────────────────────────────────────────────
...